add support for json configuration
