+++ EasyConfig.Net.Core/Attributes/EnvironmentAttribute.cs
@@ -2,7 +2,7 @@ namespace EasyConfig.Attributes
 {
     public class EnvironmentAttribute : ConfigurationAttribute
     {
-        public EnvironmentAttribute(string key, string alias = "") : base(key ,alias, ConfigurationSources.Environment)
+        public EnvironmentAttribute(string key) : base(key, "", ConfigurationSources.Environment)
         {
         }
     }
