+++ EasyConfig.Net.Core/Attributes/JsonConfig.cs
@@ -1,0 +1,9 @@
+namespace EasyConfig.Attributes
+{
+    public class JsonConfig : ConfigurationAttribute
+    {
+        public JsonConfig(string key) : base(key, "", ConfigurationSources.ConfigFile)
+        {
+        }
+    }
+}
