+++ EasyConfig.Net.Core/ConfigurationSources.cs
@@ -3,6 +3,7 @@ namespace EasyConfig
     public enum ConfigurationSources
     {
         Environment,
-        CommandLine
+        CommandLine,
+        ConfigFile
     }
 }
