+++ EasyConfig.Net.Core/EasyConfig.Net.Core.csproj
@@ -5,4 +5,9 @@
     <RootNamespace>EasyConfig</RootNamespace>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="1.1.2" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="1.1.2" />
+  </ItemGroup>
+
 </Project>
